feat(gallery): apply new specs - INNO-1162 (#952)
# PR description

add component gallery using provided [specifications](https://webgate.ec.europa.eu/CITnet/confluence/display/NEXTEUROPA/Gallery+Design+Specifications) (and also updates mades in the comments, but not visible yet on the mockups)

A few remarks concerning the display of images:
To achieve required design (always full width on each line), there is a css property `object-fit` which works just fine. But of course, it is not supported by IE. 
So display is conform to specifications on modern browsers, and partially degraded on IE.

**Important to notice:** 
this PR add the possibility to have icons before the label in Buttons (needed for the overlay).